fix: configure arquiteture for new model
