--- src/Components/Footer/index.jsx
@@ -1,14 +1,0 @@
-export default function Footer() {
-  return (
-    <footer className='bg-amber-900 w-full h-10 fixed bottom-0 flex items-center justify-center text-white'>
-      <ul className='w-full relative flex justify-around text-sm'>
-        <li><a href="https://github.com/edrikfsteiner" className='text-yellow-600' target="_blank" rel="noopener noreferrer">edrik.github</a></li>
-        <li><a href="https://github.com/ribarski" className='text-yellow-600' target="_blank" rel="noopener noreferrer">eduardo.github</a></li>
-        <li><a href="https://github.com/erycmj" className='text-yellow-600' target="_blank" rel="noopener noreferrer">eryc.github</a></li>
-        <li><a href="https://github.com/isiscostabb" className='text-yellow-600' target="_blank" rel="noopener noreferrer">isis.github</a></li>
-        <li><a href="https://github.com/miguelfermo" className='text-yellow-600' target="_blank" rel="noopener noreferrer">miguel.github</a></li>
-        <li><a href="https://github.com/taylorteixeira" className='text-yellow-600' target="_blank" rel="noopener noreferrer">taylor.github</a></li>
-      </ul>
-    </footer >
-  );
-}
