--- src/Components/HeroContent/HeroCard/HeroContentTitle/index.jsx
@@ -1,7 +1,0 @@
-export function HeroContentTitle() {
-    return (
-        <h1 className="text-center text-5xl font-bold text-orange-950 mb-8">
-            SOBRE O NOSSO SITE
-        </h1>
-    );
-}
