--- src/Components/HomeFooter/data.js
@@ -1,1 +1,0 @@
-
