--- src/Components/HomeFooter/index.jsx
@@ -1,1 +1,0 @@
-
