--- src/Components/HomeFooter/style.js
@@ -1,1 +1,0 @@
-
