--- src/context/DonationsContext.jsx
@@ -1,2 +1,0 @@
-export { default as DonationsProvider } from "@/features/donations/context/DonationsProvider"
-export { useDonations } from "@/features/donations/context/useDonations"
