--- src/context/UserContext.jsx
@@ -1,22 +1,0 @@
-// src/context/UserContext.js
-import { createContext, useState, useContext } from "react"
-import PropTypes from "prop-types"
-
-const UserContext = createContext()
-
-export const UserProvider = ({ children }) => {
-  const [user, setUser] = useState(null)
-
-  return (
-    <UserContext.Provider value={{ user, setUser }}>
-      {children}
-    </UserContext.Provider>
-  )
-}
-
-UserProvider.propTypes = {
-  children: PropTypes.node.isRequired,
-}
-
-// eslint-disable-next-line react-refresh/only-export-components
-export const useUser = () => useContext(UserContext)
