+++ src/features/donations/components/Donations.jsx
@@ -2,7 +2,7 @@ import { useState } from "react"
 import PropTypes from "prop-types"
 import { BiTimeFive } from "react-icons/bi"
 import Modal from "./Modal"
-import { useDonations } from "@/context/DonationsContext"
+import { useDonations } from "@/features/donations/hooks/useDonations"
 import Button from "@/shared/ui/Button"
 import Paragraph from "@/shared/ui/Paragraph"
 import Title from "@/shared/ui/Title"
