+++ src/features/donations/hooks/useDonations.js
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import DonationsContext from "./DonationsContext";
+import DonationsContext from "../context/DonationsContext";
 
 export function useDonations() {
     return useContext(DonationsContext);
