+++ src/features/hero/components/HeroContentButton.jsx
@@ -4,7 +4,7 @@ import { Link } from "react-router-dom";
 export function HeroContentButton() {
     return (
         <div className="flex justify-center m-4">
-            <Link to="/Donations">
+            <Link to="/donations">
                 <Button className="w-72 rounded-lg bg-orange-400 py-2 text-2xl font-bold tracking-widest text-orange-900 shadow-md transition-colors duration-200 hover:bg-orange-500 focus-visible:ring-2 focus-visible:ring-yellow-500 focus-visible:ring-offset-2">
                     NAVEGAR
                 </Button>
