+++ src/features/hero/pages/HeroPage.jsx
@@ -1,0 +1,13 @@
+import Footer from "@/features/hero/components/HeroFooter"
+import HeaderScreen from "@/features/hero/components/HeroHeader"
+import HeroContent from "@/features/hero/components/HeroContent"
+
+export default function HeroScreen() {
+    return (
+        <div>
+            <HeaderScreen />
+            <HeroContent />
+            <Footer />
+        </div>
+    );
+}
