--- src/pages/DonationManager/index.jsx
@@ -1,3 +1,0 @@
-export default function DonationManagerPage() {
-  return null
-}
