--- src/pages/Donations/index.jsx
@@ -1,3 +1,0 @@
-export default function DonationsPage() {
-  return null
-}
