--- src/pages/Hero/index.jsx
@@ -1,13 +1,0 @@
-import Footer from "@/Components/Footer/index"
-import HeaderScreen from "@/Components/Header/index"
-import { HeroContent } from "@/Components/HeroContent/index"
-
-export default function HeroScreen() {
-    return (
-        <div>
-            <HeaderScreen />
-            <HeroContent />
-            <Footer />
-        </div>
-    );
-}
