--- src/pages/Home/index.jsx
@@ -1,3 +1,0 @@
-export default function HomePage() {
-  return null
-}
