--- src/router/routes.jsx
@@ -1,1 +1,0 @@
-export { router as routes } from "../app/router"
