--- src/shared/contexts/Donations/index.jsx
@@ -1,3 +1,0 @@
-/* eslint-disable react-refresh/only-export-components */
-export { default as DonationsProvider } from "@/features/donations/context/DonationsProvider"
-export { useDonations } from "@/features/donations/context/useDonations"
