--- src/shared/contexts/User/index.jsx
@@ -1,2 +1,0 @@
-/* eslint-disable react-refresh/only-export-components */
-export { UserProvider, useUser } from "../../../context/UserContext"
